refactor: relocate mobile constants to a dedicated shared directory
