+++ src/apps/content/composables/useContentAppUIState.js
@@ -1,6 +1,6 @@
 import { ref, computed, onMounted } from 'vue';
 import { deviceDetector } from '@/utils/browser/compatibility.js';
-import { MOBILE_CONSTANTS } from '@/shared/config/constants.js';
+import { MOBILE_CONSTANTS } from '@/shared/constants/mobile.js';
 import { getScopedLogger } from '@/shared/logging/logger.js';
 import { LOG_COMPONENTS } from '@/shared/logging/logConstants.js';
